full fledged curd operation codes done

+++ src/main/java/com/products/listing/repository/ProductRepository.java
@@ -4,6 +4,8 @@ import com.products.listing.model.ProductDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface ProductRepository extends JpaRepository<ProductDetails, Integer> {
+public interface ProductRepository extends JpaRepository<ProductDetails, Long> {
 }
